1. move examples under root directory. 2. rename error.rs 3. add README.md
