+++ README.md
@@ -1,0 +1,11 @@
+##See example
+```
+start_miner()
+
+start_server()
+```
+
+##Run example
+`
+cargo run --release --example connect
+`
